+++ CHANGELOG.md
@@ -170,3 +170,19 @@ can get data for both organization and organization brand pages.
 
 ## 4.3.0 (Mar 8, 2023)
 * Add support to upload video files from URLs
+
+## 4.3.1 (Mar 14, 2023)
+* Add a flag to note that we want to upload a thumbnail image along with the video
+
+## 4.4.0 (Mar 16, 2023)
+* Add support to upload thumbnail image to a video
+
+## 4.4.1 (Mar 20, 2023)
+* Add more detail to debug logs
+* Add support to retrieve video details from LinkedIn
+* Extend `Post` object to include video details mapping
+
+## 4.4.2 (Mar 22, 2023)
+* Add support to retrieve video details from LinkedIn
+
+## 4.4.3 (Work in progress)